waypoint-backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
anthropic>=0.28.0
pdfplumber==0.11.0
python-docx==1.1.2
networkx==3.3
python-multipart==0.0.9
pydantic==2.7.1
spacy==3.8.3
python-dotenv==1.0.1
